+++ Assets/_Project/Scripts/Player/Movement/MovementState.cs
@@ -1,0 +1,12 @@
+///<summary>
+/// Represents the players current locomotive state
+/// </summary>
+
+public enum MovementState
+{
+    Idle,
+    Walking,
+    Running,
+    Jumping,
+    Falling
+}